+++ changelog.md
@@ -16,6 +16,8 @@ Codion Change Log
 - Column.GetValue and SetParameter moved to common.db as GetValue and SetValue.
 - ColumnDefinition.LEGACY_JDBC moved to Database.
 - MultiValueCondition, upper(val) now uppercase.
+### is.codion.framework.json.domain
+- EntityDeserializer now respects Entities.STRICT_DESERIALIZATION.
 ### is.codion.framework.servlet
 - EntityService.SERIALIZATION config param typo fixed.
 ### is.codion.framework.model